+++ packages/jest-transform/src/types.ts
@@ -31,10 +31,10 @@ interface FixedRawSourceMap extends Omit<RawSourceMap, 'version'> {
   version: number;
 }
 
-// TODO: For Jest 26 normalize this (always structured data, never a string)
-export type TransformedSource =
-  | {code: string; map?: FixedRawSourceMap | string | null}
-  | string;
+export type TransformedSource = {
+  code: string;
+  map?: FixedRawSourceMap | string | null;
+};
 
 export type TransformResult = TransformTypes.TransformResult;